backend1/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-environ
django-filter
Pillow
psycopg2-binary
whitenoise[brotli]
django-cors-headers
djoser
gunicorn